mkdir is log folder doesn't exist
The original intention was for the user to input the directory they already use for logs to put the tsd logs in the same folder. Realized the user may not know where the logs are currently and will go with the default. Changing to check for the inputted folder and create it if it doesn't exist.