Drop existing handlers in setup_logging
Calls to setup_logging always created and attached a new set of handlers.

When waiting for a working auth_token, current logic keeps reconfiguring
logging, according to the actual content of the cfg file.

It would be possible to extract logging setup out of that loop, but I
went with this solution: dropping existing handlers when setting up root
handler.